feat: enum체크할 커스텀 어노테이션 추가 (#27)
