+++ module-core/src/main/java/com/back2basics/infra/post/custom/CustomEnumCheck.java
@@ -1,4 +1,4 @@
-package com.back2basics.infra.custom;
+package com.back2basics.infra.post.custom;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -9,14 +9,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Documented
-@Constraint(validatedBy = CustomNotBlankValidator.class)
+@Constraint(validatedBy = CustomEnumCheckValidator.class)
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CustomTitleNotBlank {
+public @interface CustomEnumCheck {
 
-    String message();
+    String message() default "유효하지 않은 값입니다.";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
+
+    Class<? extends Enum<?>> enumClass(); // 체크할 enum type
 }
